+++ rdsa_utils/validation.py
@@ -2,11 +2,12 @@
 import json
 import logging
 from typing import Any, Callable, Mapping, Optional
+import warnings
 
 import pandas as pd
 from pydantic import BaseModel, validator
 
-from rdsa_utils.helpers.helpers_python import list_convert
+from rdsa_utils.helpers.python import list_convert
 
 
 logger = logging.getLogger(__name__)
@@ -35,7 +36,9 @@ def apply_validation(
         The input config after being passed through the validator.
     """
     if not Validator:
-        logger.warning('No validator provided, config contents unvalidated.')
+        msg = 'No validator provided, config contents unvalidated.'
+        logger.warning(msg)
+        warnings.warn(msg, stacklevel=2)
         return config
 
     validated_config = Validator(**config).model_dump(exclude_unset=True)